Add CLoader, CDumper, refactored Loaders, Dumpers
