+++ ci/run_tests.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
 
 #- coverage run -m unittest discover
-coverage run  --branch -m unittest discover
-coverage report yamlloader/*.py
+coverage run  --branch -m unittest discover && coverage report yamlloader/*.py
