+++ yamlloader/ordereddict/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+"""YAML loaders and dumpers for PyYAML allowing to keep keys order."""
+from __future__ import print_function, division, absolute_import
+
+from .loaders import Loader, SafeLoader, CLoader, CSafeLoader
+from .dumpers import Dumper, SafeDumper, CDumper, CSafeDumper
+
+__all__ = ['CLoader',
+           'Loader',
+           'CDumper',
+           'Dumper',
+           'CSafeLoader',
+           'SafeLoader',
+           'CSafeDumper',
+           'SafeDumper']
